languages: add python support
For builds outside the sandbox, the interpreter that was used to execute
bob is used. Inside a sandbox, a python executable must be in $PATH. In
any case, the interpreter is started to ignore user installed modules.
This shall ensure that a plain interpreter is used and any additional
modules need to be supplied by the recipes. Hopefully, this is enough to
prevent too much host state to creep into the build.

By default, the "os", "os.path" and "sys" modules are imported. They are
needed by the Bob script prologue and will be required by any script
anyway.

Environment variables are set in the os.environ dict. The various
standard mapping types (BOB_ALL_PATHS, BOB_DEP_PATHS and BOB_TOOL_PATHS)
are declared as standard python dictionaries as global variables.

Other than that, the integration is just adding the extensions to the
various "script" keywords.